Avances de Inmuebel terminado

+++ app/src/main/java/com/example/inmobiliaria_2025/model/Inmueble.java
@@ -13,7 +13,7 @@ public class Inmueble implements Serializable {
 
     private int ambientes;
 
-    private double superficie;
+    private int superficie;
     private double latitud;
     private double longitud;
 
@@ -24,7 +24,7 @@ public class Inmueble implements Serializable {
     public Inmueble() {
     }
 
-    public Inmueble(int idInmueble, String direccion, String tipo, String uso, double valor, Propietario propietario, int ambientes, double superficie, double latitud, double longitud, String imagen, boolean disponible, Propietario duenio) {
+    public Inmueble(int idInmueble, String direccion, String tipo, String uso, double valor, Propietario propietario, int ambientes, int superficie, double latitud, double longitud, String imagen, boolean disponible, Propietario duenio) {
         this.idInmueble = idInmueble;
         this.direccion = direccion;
         this.tipo = tipo;
@@ -100,7 +100,7 @@ public class Inmueble implements Serializable {
         return superficie;
     }
 
-    public void setSuperficie(double superficie) {
+    public void setSuperficie(int superficie) {
         this.superficie = superficie;
     }
